+++ tests/io/UploadDescriptorFileReaderTestCase.py
@@ -1,13 +1,13 @@
 from sdk_entrepot_gpf.Errors import GpfSdkError
 from sdk_entrepot_gpf.io.Config import Config
 from sdk_entrepot_gpf.io.Dataset import Dataset
-from sdk_entrepot_gpf.io.DescriptorFileReader import DescriptorFileReader
+from sdk_entrepot_gpf.io.UploadDescriptorFileReader import UploadDescriptorFileReader
 
 from tests.GpfTestCase import GpfTestCase
 
 
-class DescriptorFileReaderTestCase(GpfTestCase):
-    """Test de la classe DescriptorFileReader.
+class UploadDescriptorFileReaderTestCase(GpfTestCase):
+    """Test de la classe UploadDescriptorFileReader.
 
     cmd : python3 -m unittest -b tests.io.DescriptorFileReaderTestCase
     """
@@ -15,7 +15,7 @@ class DescriptorFileReaderTestCase(GpfTestCase):
     def test_init_ok_1(self) -> None:
         """Test du constructeur quand tout va bien n°1."""
         # Ouverture
-        o_dsr = DescriptorFileReader(Config.data_dir_path / "datasets" / "1_dataset_vector" / "upload_descriptor.json")
+        o_dsr = UploadDescriptorFileReader(Config.data_dir_path / "datasets" / "1_dataset_vector" / "upload_descriptor.json")
         # Vérifications
         self.assertEqual(len(o_dsr.datasets), 1)
         self.assertIsInstance(o_dsr.datasets[0], Dataset)
@@ -23,7 +23,7 @@ class DescriptorFileReaderTestCase(GpfTestCase):
     def test_init_ok_2(self) -> None:
         """Test du constructeur quand tout va bien n°2."""
         # Ouverture
-        o_dsr = DescriptorFileReader(Config.data_dir_path / "datasets" / "2_dataset_archive" / "upload_descriptor.json")
+        o_dsr = UploadDescriptorFileReader(Config.data_dir_path / "datasets" / "2_dataset_archive" / "upload_descriptor.json")
         # Vérifications
         self.assertEqual(len(o_dsr.datasets), 1)
         self.assertEqual(o_dsr.datasets[0].upload_infos["name"], "EXAMPLE_DATASET_ARCHIVE")
@@ -32,6 +32,6 @@ class DescriptorFileReaderTestCase(GpfTestCase):
         """Test du constructeur quand au moins un dossier indiqué n'est pas trouvé."""
         with self.assertRaises(GpfSdkError) as o_arc:
             # Ouverture
-            DescriptorFileReader(GpfTestCase.data_dir_path / "datasets" / "1_test_dataset_bad_pathes" / "upload_descriptor.json")
+            UploadDescriptorFileReader(GpfTestCase.data_dir_path / "datasets" / "1_test_dataset_bad_pathes" / "upload_descriptor.json")
         # Vérifications
         self.assertEqual(o_arc.exception.message, "Au moins un des répertoires listés dans le fichier descripteur de livraison n'existe pas.")